refactor(pr-reviewer): split PRReviewer into modular sub-components
